refactor: move automated-lead to top-level command
Move automated-lead from an alert subcommand to a top-level command
to separate falcon-cli extended commands (multi-API) from Swagger-
generated REST resource commands. The Command enum and help text
mark it as [multi-API] to indicate it does not map 1:1 to a single
Falcon API endpoint.
